+++ ArcBlockSDK.h
@@ -1,4 +1,0 @@
-#import <ABSDKCoreKit/ABSDKCoreKit.h>
-#import <ABSDKAccountKit/ABSDKAccountKit.h>
-#import <ABSDKMessagingKit/ABSDKMessagingKit.h>
-#import <ABSDKWalletKit/ABSDKWalletKit.h>